Refactor WS
